pyproject.toml
[project]
name = "zerobs"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.6.2",
    "docling>=2.32.0",
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
